% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainline_support_functions.R
\name{as_boolean}
\alias{as_boolean}
\title{Convert NULL values to FALSE and non-NULLs to their logical value}
\usage{
as_boolean(value)
}
\arguments{
\item{value}{value to be converted to boolean}
}
\value{
input value as a boolean
}
\description{
The core function as.logical() crashes if you give it a NULL.
This convenience function turns those NULLs into FALSE and preserves
the same output as as.logical() for non-NULLs.
}
\details{
It probably needs a bit more work to make it behave the same way as
as.logical() does for vectors.  For the moment, this is just working
on single values when they are NULL.
}
