% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainline_support_functions.R
\name{clean_up_after_tzar_emulation}
\alias{clean_up_after_tzar_emulation}
\title{Get rid of any traces of running tzar emulation if emulating.}
\usage{
clean_up_after_tzar_emulation(tzar_in_progress_dir_name,
  tzar_emulation_completed_dir_name, copy_model_dot_R_tzar_file,
  full_model_dot_R_DEST_path, emulation_scratch_file_path,
  echo_console_to_temp_file, console_sink_file_info, full_output_dir_with_slash)
}
\description{
Need to rename the output directory to something to do
 with emulation so that if you go back to a pile of directories
 and find runs with strange output, you know that it's because
 you were playing around with the code in the emulator.
 Otherwise, you might have partial results in a run that
 appeared to have run to completion because tzar DID run to
 completion on the dummy model.R code that enabled the
 emulation to build the tzar directories and parameters.
}
\details{
Can't do this earlier because you don't know how many
 things inside the parameters list have the inprogress name
 built into them by tzar.  If you changed the inprogress
 directory name before running the user code, then the run
 would be looking for the wrong directory.

However, if tzar itself knew that this was going to be an
 emulation name, it could just use the emulation name wherever
 it now uses the in progress name and it could skip the renaming
 after the run is complete.  If those two things were happening,
 then this cleanup code would be unnecessary.
}
