% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console_sink_file_functions.R
\name{clean_up_console_sink}
\alias{clean_up_console_sink}
\title{Clean up the sink file created for console output if necessary}
\usage{
clean_up_console_sink(echo_console_to_temp_file, console_sink_file_info,
  full_output_dir_with_slash)
}
\arguments{
\item{echo_console_to_temp_file}{boolean}

\item{console_sink_file_info}{list containing named elements
console_output_file_name, sink_file_path, temp_console_out_file}

\item{full_output_dir_with_slash}{file}
}
\value{
Returns nothing
}
\description{
Clean up the sink file created for console output if necessary
}
