% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console_sink_file_functions.R
\name{clean_up_console_sink_if_necessary}
\alias{clean_up_console_sink_if_necessary}
\title{Clean up the sink file created for console output if requested}
\usage{

  clean_up_console_sink_if_necessary(tzar_emulation_flag_and_console_sink_information)
}
\arguments{
\item{tzar_emulation_flag_and_console_sink_information}{List containing
emulating_tzar flag, echo_console_to_temp_file flag, and
temp_console_out_file.}
}
\value{
Returns nothing
}
\description{
Clean up the sink file created for console output if requested
}
