% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console_sink_file_functions.R
\name{get_tzar_emulation_flag_and_console_sink_if_requested}
\alias{get_tzar_emulation_flag_and_console_sink_if_requested}
\title{Get value for tzar emulation flag and create console sink file if asked}
\usage{
get_tzar_emulation_flag_and_console_sink_if_requested(parameters)
}
\arguments{
\item{parameters}{List of parameters read from project.yaml file.}
}
\value{
Returns List containing emulating_tzar flag,
echo_console_to_temp_file flag, and temp_console_out_file.
}
\description{
Get value for tzar emulation flag and create console sink file if asked
}
