% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console_sink_file_functions.R
\name{open_sink_file}
\alias{open_sink_file}
\title{Open console sink file for saving all console output during a run}
\usage{
open_sink_file(console_out_file_name_with_path)
}
\arguments{
\item{sink_file_path}{tzar_em_scratch_dir Location of tzar emulation scratch
directory as path string with no trailing slash}

\item{console_output_file_name}{console_output_file_name file name string for file to
contain console sink output without any path information in the string}
}
\value{
Returns console_sink_file_info list containing named elements
console_output_file_name, sink_file_path, temp_console_out_file
}
\description{
Open console sink file for saving all console output during a run
}
