% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console_sink_file_functions.R
\name{open_sink_file_if_requested}
\alias{open_sink_file_if_requested}
\title{Open console sink file for saving all console output during a run if asked}
\usage{
open_sink_file_if_requested(emulating_tzar, echo_console_to_temp_file,
  full_output_dir_with_slash,
  console_output_file_name = "console_sink_output.temp.txt")
}
\arguments{
\item{emulating_tzar}{boolean}

\item{echo_console_to_temp_file}{boolean}

\item{full_output_dir_with_slash}{path string}

\item{console_output_file_name}{file name string}
}
\value{
Returns List containing emulating_tzar flag,
echo_console_to_temp_file flag, and temp_console_out_file.
}
\description{
Open console sink file for saving all console output during a run if asked
}
