% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulateRunningUnderTzar.R
\name{run_mainline_under_tzar_or_tzar_emulation}
\alias{run_mainline_under_tzar_or_tzar_emulation}
\title{Run a function under normal tzar or tzar emulation}
\usage{
run_mainline_under_tzar_or_tzar_emulation(main_function, projectPath,
  tzarJarPath, emulation_scratch_file_path = "~/tzar_emulation_scratch.yaml",
  emulatingTzar = TRUE)
}
\arguments{
\item{main_function}{function to call to run under tzar or tzar emulation
(NOTE: NOT a string)}

\item{projectPath}{path of R code and project.yaml file for project}

\item{tzarJarPath}{Path to the jar file to use to run tzar}

\item{emulation_scratch_file_path}{path of scratch file for passing
tzarEmulation flag and tzarOutputDir between tzar and mainline function}

\item{emulatingTzar}{boolean with TRUE indicating main_function should be
run under tzar emulation and FALSE indicating run under normal tzar.}
}
\value{
parameters list of parameters loaded from project.yaml file
}
\description{
Run a function under normal tzar or tzar emulation
}
\examples{
\dontrun{
run_mainline_under_tzar_or_tzar_emulation (main_function=trial_main_function,
         projectPath=".",
         tzarJarPath = "~/D/rdv-framework-latest-work/tzar.jar",
         emulation_scratch_file_path="~/tzar_emulation_scratch.yaml",
         emulatingTzar=TRUE
         )
}
}

