% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulateRunningUnderTzar.R
\name{run_tzar}
\alias{run_tzar}
\title{Run a function under normal tzar or tzar emulation}
\usage{
run_tzar(emulating_tzar = emulating_tzar, main_function = main,
  project_path = project_path,
  emulation_scratch_file_path = emulation_scratch_file_path,
  tzar_jar_path = tzar_jar_path, copy_model_R_tzar_file = FALSE,
  model_R_tzar_src_dir = model_R_tzar_src_dir,
  model_R_tzar_disguised_filename = "model.R.tzar",
  overwrite_existing_model_R_dest = TRUE,
  required_model_R_filename_for_tzar = "model.R")
}
\arguments{
\item{emulating_tzar}{boolean with TRUE indicating main_function should be
run under tzar emulation and FALSE indicating run under normal tzar}

\item{main_function}{function to call to run under tzar or tzar emulation
(NOTE: NOT a string), i.e., your main application code}

\item{project_path}{path of R code and project.yaml file for project}

\item{emulation_scratch_file_path}{path of scratch file for passing
tzarEmulation flag and tzarOutputDir between tzar and mainline function}

\item{tzar_jar_path}{Path to the jar file to use to run tzar}

\item{copy_model_R_tzar_file}{Boolean flag indicating whether model.R must
be copied in (e.g., when running inside a package)}

\item{model_R_tzar_src_dir}{Path to directory holding the model.R file to
use to run tzar}

\item{model_R_tzar_disguised_filename}{Name (without path) of the file
containing the code to be run by tzar as model.R}

\item{overwrite_existing_model_R_dest}{Boolean flag indicating whether
model.R being copied in should overwrite any existing model.R in destination}

\item{required_model_R_filename_for_tzar}{Name of file that tzar expects to
find and execute to call user's application code from tzar (currently,
it's always "model.R")}
}
\value{
parameters list of parameters loaded from project.yaml file
}
\description{
Run a function under normal tzar or tzar emulation
}
\examples{
\dontrun{
run_tzar (
         emulating_tzar              = TRUE,
         main_function               = trial_main_function,
         project_path                = ".",
         emulation_scratch_file_path = "~/tzar_emulation_scratch.yaml",
         tzar_jar_path               = "~/D/rdv-framework-latest-work/tzar.jar",
         copy_model_R_tzar_file             = FALSE,
         model_R_tzar_src_dir               = model_R_tzar_src_dir,
         model_R_tzar_disguised_filename    = "model.R.tzar",
         overwrite_existing_model_R_dest    = TRUE,
         required_model_R_filename_for_tzar = "model.R"
         )
}
}

