% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainline_support_functions.R
\name{run_tzar_java_jar}
\alias{run_tzar_java_jar}
\title{Run the tzar jar under java}
\usage{
run_tzar_java_jar(tzar_jar_path, project_path)
}
\arguments{
\item{tzar_jar_path}{Path of directory where tzar jar is found}

\item{project_path}{Path of directory where user code to be run is found}
}
\description{
Run the tzar jar under java
}
