% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulateRunningUnderTzar.R
\name{run_tzar_outside_pkg}
\alias{run_tzar_outside_pkg}
\title{Convenience function to call run_tzar() when code is not part of a package}
\usage{
run_tzar_outside_pkg(emulating_tzar, main_function, project_path,
  emulation_scratch_file_path, tzar_jar_path)
}
\arguments{
\item{emulating_tzar}{boolean with TRUE indicating main_function should be
run under tzar emulation and FALSE indicating run under normal tzar}

\item{main_function}{function to call to run under tzar or tzar emulation
(NOTE: NOT a string), i.e., your main application code}

\item{project_path}{path of R code and project.yaml file for project}

\item{emulation_scratch_file_path}{path of scratch file for passing
tzarEmulation flag and tzarOutputDir between tzar and mainline function}

\item{tzar_jar_path}{Path to the jar file to use to run tzar}
}
\value{
parameters list of parameters loaded from project.yaml file
}
\description{
Convenience function to call run_tzar() when code is not part of a package
}
\examples{
\dontrun{
run_tzar (
         emulating_tzar              = TRUE,
         main_function               = trial_main_function,
         project_path                = ".",
         emulation_scratch_file_path = "~/tzar_emulation_scratch.yaml",
         tzar_jar_path               = "~/D/rdv-framework-latest-work/tzar.jar"
         )
}
}

